feat(fga): align ListPermissions with expanded schema
- split fga.go into check_permissions.go / list_permissions.go to
  match the one-file-per-API layout
- ListPermissions: relation/object_type now optional (empty request
  lists everything the caller holds); response adds permissions
  (object/relation detail) and truncated (1000-entry cap flag)
- widen FGA-unavailable skip detection so integration tests skip
  cleanly against pre-FGA server images
